moved and refactored all components to options page
