+++ client/src/App.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import { Home, Options, Disclaimer, NotFound } from './components/layouts'
-import { UserForm } from './components/UserForm'
+
 import './App.css';
-import { WeeklyCharts } from './components/WeeklyCharts';
+
+ 
 
 class App extends React.Component {
   render() {
@@ -14,8 +15,6 @@ class App extends React.Component {
           <Route path="/disclaimer" component={Disclaimer} />
           <Route path="/options" component={Options} />
           <Route component={NotFound} />
-          <Route component={UserForm} />
-          <Route component={WeeklyCharts} />
         </Switch>
       </BrowserRouter>
