--- client/src/components/Grid/index.js
@@ -1,3 +1,0 @@
-export * from "./Col";
-export * from "./Container";
-export * from "./Row";
