--- client/src/components/UserForm/index.js
@@ -1,1 +1,0 @@
-export {default as UserForm} from "./UserForm";
