--- client/src/components/form/index.js
@@ -1,2 +1,0 @@
-export * from "./input";
-export * from "./FormBtn";
