+++ client/src/components/layouts/Options/components/Col/index.js
@@ -1,7 +1,11 @@
 import React from "react";
 
-export const Col = ({ size, children }) => (
+   const Col = ({size, children}) =>{
+     return ( 
   <div className={size.split(" ").map(size => "col-" + size).join(" ")}>
     {children}
   </div>
 );
+}
+
+export default Col;
