+++ client/src/components/layouts/Options/components/Container/index.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+const Container = ({fluid, children}) => {
+  return (
+    <div className={`container${fluid ? 'fluid' : ''}`}>
+      {children}
+      </div>
+  )
+}
+
+
+export default Container;
