+++ client/src/components/layouts/Options/components/Row/index.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+const Row = ({ fluid, children }) => {
+  return (
+    <div className={`row${fluid ? "-fluid" : ""}`}>
+      {children}
+    </div>)
+}
+
+export default Row;
