+++ client/src/components/layouts/Options/components/UserForm/components/Button.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+const Button = props => {
+ return (
+  <button 
+  style={{ float: "left", marginBottom: 10 }} 
+  className={`btn ${props.btnColor}`}
+  >
+  
+  {props.children}
+  </button>
+   
+);
+}
+
+export default Button;
