+++ client/src/components/layouts/Options/components/UserForm/components/Input.js
@@ -1,7 +1,11 @@
 import React from "react";
 
-export const Input = props => (
+ const Input = props => { 
+ return (
   <div className="form-group">
     <input className="form-control" {...props} />
   </div>
-);
+)
+ }
+
+ export default Input
