+++ client/src/components/layouts/Options/components/UserForm/components/index.js
@@ -1,0 +1,4 @@
+export {default as Button} from './Button';
+export {default as Input} from './Input';
+ 
+
