+++ client/src/components/layouts/Options/components/WeeklyCharts/components/CalorieBarChart/index.jsx
