+++ client/src/components/layouts/Options/components/WeeklyCharts/components/index.js
@@ -1,0 +1,2 @@
+export {default as CalorieBarChart} from './CalorieBarChart';
+export {default as MacroPieChart} from './MacroPieChart';
