+++ client/src/components/layouts/Options/components/index.jsx
@@ -1,0 +1,7 @@
+ 
+export {default as UserForm} from "./UserForm";
+export {default as WeeklyCharts} from "./WeeklyCharts";
+export {default as Col} from "./Col";
+export {default as Container} from "./Container";
+export {default as Row} from "./Row";
+
